requirements.txt
absl-py==1.0.0
adal==1.2.7
altgraph==0.17.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.4.0
auditlog3==1.0.1
auto-py-to-exe==2.16.0
autopep8==1.6.0
awscli==1.22.32
azure-core==1.19.0
azure-storage-blob==12.9.0
backcall==0.2.0
bleach==4.1.0
boto3==1.6.23
botocore==1.9.23
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==5.0.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
customtkinter==3.8
cycler==0.11.0
darkdetect==0.5.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.2
django-appconf==1.0.2
django-cors-headers==3.10.0
django-filter==21.1
django-mongodb-engine-py3==0.6.0.1
django-multiselectfield==0.1.8
django-rest-swagger==2.2.0
django-staticfiles==1.2.1
django-storages==1.6.6
djangorestframework==3.12.4
djangotoolbox==1.8.0
dnspython==2.1.0
docutils==0.14
EasyProcess==1.1
Eel==0.12.4
entrypoint2==1.0
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.6.0
Flask==2.0.3
flatbuffers==2.0
fonttools==4.28.5
fpdf==1.7.2
future==0.18.2
gast==0.4.0
gevent==21.12.0
gevent-websocket==0.10.1
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
googletrans==3.0.0
greenlet==1.1.2
grpcio==1.43.0
gunicorn==19.7.1
h11==0.9.0
h2==3.2.0
h5py==3.6.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.20.4
httpx==0.13.3
hubspot3==3.2.51
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.8.1
ipykernel==6.6.1
ipython==7.30.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.1
jmespath==0.9.3
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
labelImg==1.8.6
libclang==12.0.0
libretranslatepy==2.1.1
lml==0.1.0
lxml==4.7.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mongoengine==0.23.1
MouseInfo==0.1.3
msrest==0.6.21
msrestazure==0.6.4
mss==6.1.0
mysqlclient==2.1.0
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.22.0
oauth2client==4.1.3
oauthlib==3.1.1
openapi-codec==1.3.2
opencv-python==4.5.5.62
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pefile==2021.9.3
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.20
PyDrive==1.3.1
pyexcel-io==0.6.5
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
PyGetWindow==0.0.9
pyglet==1.5.23
Pygments==2.11.1
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.0
PyJWT==2.2.0
pymongo==3.12.1
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.1.4
pyrsistent==0.18.0
pyscreenshot==2.3
PyScreeze==0.1.28
pystray==0.19.2
python-dateutil==2.6.1
python-decouple==3.3
python-http-client==2.2.1
python-snappy==0.6.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2018.3
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.8
s3transfer==0.1.13
schedule==1.1.0
Send2Trash==1.8.0
simplejson==3.17.5
six==1.16.0
sniffio==1.2.0
SpeechRecognition==3.8.1
sqlparse==0.4.2
static3==0.7.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tk==0.1.0
toml==0.10.2
torch==1.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
translate==3.6.1
typing-extensions==3.10.0.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.6
vine==1.1.4
virtualenv==20.13.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whichcraft==0.6.1
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
